Add product component
Much of this markup really shouldn't be here. When we've got some
functionality to speak of, start stripping the markup down into
its own separated theme, which will rely on this plugin. That way
people can either install the theme for a fully functioning site,
or just install the Shop plugin and write the markup themselves.