We're in business! Well, nearly...
Adds composer.json and register facade/service provider to get the
Cart package running, and also adds a basket component to list the
contents of a user's basket. Last but not least, strips out them
ridiculous +/- buttons from the qty box and adds Ajax magic to the
product forms such that things can be added to the basket.