Check category exists before loading products
The demo theme is using the category page as the page that lists
products as well, since it makes no sense to *just* have a page
that lists the categories like I did in dev. This extra bit of code
makes sure a category has been passed in before trying to load any
products to prevent errors.

This whole one-page-deal should nicely open the doors for filters too.