Fix issue with component not accepting params
When creating two instances of the Basket partial on a single page,
eg one in the header and one for the checkout view, in-line parameters
were not working. Add an onRender method that allows for these values
to be parsed, and separate parameters based on when they run.

Also adds two new parameters for specifying the basket component and
partial, which helps to fix some other issues I had with updating the
minibasket on the demo theme.