Add images, finally closes #1
Basically adds all the stuff for product images in the plugin. Removes
the old product container class, instead using the row class that's also
used on other components. Adds a few new properties for image sizes and
classes as well.

Also adds the necessary stuff to get the uploader running in the backend
product editor and tweaks the component view to display mutiple images.