Add ability to remove products from basket
Adds a new method that removes a given row from the cart to begin implementing
task #4. Also updates the default basket component to include a remove button
which will remove that row and update basket total and item count with the aid
of new ocshop-x helper classes, where 'x' is 'total', 'count' or 'row-{ROWID}'.

Finally rejigs the prepareVars method to clean things up, since much of
it is used in the onFoo listeners.