Add frontend invoice
After payment there's been no way of knowing whether the order was
successful. This adds an order component in order for order details
to be displayed on successful checkout.

Also updates the model to de-jsonify the items and convert it into
an array so that items can be looped within partials.