Use product image in category pages
Updates the products component to show the real image, if available.

TODO: Figure out a way of doing this without the horrible for loop
workaround in the view. Simply grabbing the first() image will not
work, so this could be a bug in October. Dumping $image inside the
getSquareThumbnail method of the model returns identical data as
you'd get when calling from within a loop, yet getThumb() throws an
error saying it's a non-object. Ahwel. This works for now.

In future, would be nice to set a custom "no image" image, change
thumb sizes, etc..