Add address sidebar to backend orders
Only the first half of the job, but this adds a read-only copy-pastable
version of the billing and shipping addresses to the right-hand side of
the order update view in the backend. Haven't removed the tabs yet as that
would prevent users from editing addresses, but these will need to be
replaced eventually by modal dialogues with forms inside instead.