Add address editing modal (closes #14)
Adds a bootstrap modal with dummy form with data set based on
whether it was opened from the billing or shipping address.

Also hides the tabs and adds 300px padding below the order items.
This isn't exactly pretty, but it does fix the issue where the sidebar
is really small when there's only an item or two in the order.

Long term, we probably want to set address fields to hidden, rather
than hackishly hiding the tabs they're assigned to instead...